# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/leagues/{leagueId}/image/from-url": {
      "post": {
        "operationId": "LeagueImageController_fromUrl",
        "parameters": [
          {
            "name": "leagueId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageUrlDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/LeagueImageResponseDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Gecersiz URL"
          },
          "404": {
            "description": "League not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "URL'den lig logosu",
        "tags": [
          "admin-leagues"
        ]
      }
    },
    "/api/teams": {
      "get": {
        "operationId": "TeamsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TeamResponseDto"
                      }
                    },
                    "page": {
                      "type": "number"
                    },
                    "pageSize": {
                      "type": "number"
                    },
                    "totalCount": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "pageSize",
                    "totalCount",
                    "totalPages"
                  ]
                }
              }
            }
          }
        },
        "summary": "Takim listele (paged)",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/teams/by-league/{leagueId}": {
      "get": {
        "operationId": "TeamsController_findByLeague",
        "parameters": [
          {
            "name": "leagueId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TeamResponseDto"
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Lige gore takimlar",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/teams/{id}/detail": {
      "get": {
        "operationId": "TeamsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TeamDetailDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Takim detayi (kadro + son transferler)",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/teams/{teamId}/incoming-transfers": {
      "get": {
        "operationId": "TeamsController_incoming",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TeamTransferLineDto"
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Takima gelen transferler",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/teams/{teamId}/outgoing-transfers": {
      "get": {
        "operationId": "TeamsController_outgoing",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TeamTransferLineDto"
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Takimdan giden transferler",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/teams/{teamId}/transfers": {
      "get": {
        "operationId": "TeamsController_allTransfers",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TeamTransferLineDto"
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Takimin tum transferleri (gelen+giden)",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/teams/{id}": {
      "get": {
        "operationId": "TeamsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TeamResponseDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Takimi getir",
        "tags": [
          "teams"
        ]
      }
    },
...
    "/api/admin/teams/{teamId}/image": {
      "post": {
        "operationId": "TeamImageController_upload",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TeamImageUrlResponseDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Takim logosu yukle",
        "tags": [
          "admin-teams"
        ]
      },
      "put": {
        "operationId": "TeamImageController_replace",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TeamImageUrlResponseDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Takim logosu degistir",
        "tags": [
          "admin-teams"
        ]
      },
      "delete": {
        "operationId": "TeamImageController_remove",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Takim logosu sil",
        "tags": [
          "admin-teams"
        ]
      }
    },
    "/api/admin/teams/{teamId}/image/from-url": {
      "post": {
        "operationId": "TeamImageController_fromUrl",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TeamImageUrlResponseDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "URL den takim logosu",
        "tags": [
          "admin-teams"
        ]
      }
    },
...
    "/api/admin/players/{playerId}/image/from-url": {
      "post": {
        "operationId": "PlayerImageController_fromUrl",
        "parameters": [
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ImageUploadResponseDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "URL’den oyuncu fotoğrafı",
        "tags": [
          "admin-players"
        ]
      }
    },
...
    "/api/admin/news/{newsId}/image/from-url": {
      "post": {
        "operationId": "NewsImageController_fromUrl",
        "parameters": [
          {
            "name": "newsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/NewsImageUrlResponseDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "URL den haber gorseli",
        "tags": [
          "admin-news"
        ]
      }
    },
...
  },
  "info": {
    "title": "TransferPulse API",
    "description": "TransferPulse backend (.NET → NestJS migration)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
...
      "TeamTransferLineDto": {
        "type": "object",
        "properties": {
          "transferId": {
            "type": "string"
          },
          "playerId": {
            "type": "string"
          },
          "playerName": {
            "type": "string"
          },
          "playerPhoto": {
            "type": "string"
          },
          "fromTeamId": {
            "type": "string"
          },
          "fromTeamName": {
            "type": "string"
          },
          "fromTeamLogo": {
            "type": "string"
          },
          "toTeamId": {
            "type": "string"
          },
          "toTeamName": {
            "type": "string"
          },
          "toTeamLogo": {
            "type": "string"
          },
          "transferDate": {
            "format": "date-time",
            "type": "string"
          },
          "feeAmount": {
            "type": "number"
          },
          "feeCurrency": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "transferId",
          "playerId",
          "playerName",
          "fromTeamId",
          "toTeamId",
          "transferDate",
          "feeAmount",
          "feeCurrency",
          "createdAt"
        ]
      },
...
      "LeagueImageResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://cdn.example.com/leagues/abc.png"
          }
        },
        "required": [
          "url"
        ]
      },
      "ImageUrlDto": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string",
            "description": "İndirilecek görsel URL"
          }
        },
        "required": [
          "imageUrl"
        ]
      },
      "TeamResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "leagueId": {
            "type": "string"
          },
          "leagueName": {
            "type": "string"
          },
          "playerCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "leagueId",
          "leagueName",
          "playerCount"
        ]
      },
      "SquadPlayerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "positionName": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          },
          "isFree": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "fullName",
          "nationality",
          "isFree"
        ]
      },
      "TeamDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "founded": {
            "type": "number"
          },
          "venueName": {
            "type": "string"
          },
          "venueCity": {
            "type": "string"
          },
          "venueCapacity": {
            "type": "number"
          },
          "leagueId": {
            "type": "string"
          },
          "leagueName": {
            "type": "string"
          },
          "leagueLogo": {
            "type": "string"
          },
          "playerCount": {
            "type": "number"
          },
          "squad": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SquadPlayerDto"
            }
          },
          "recentIncoming": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamTransferLineDto"
            }
          },
          "recentOutgoing": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamTransferLineDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "leagueId",
          "leagueName",
          "playerCount",
          "squad",
          "recentIncoming",
          "recentOutgoing"
        ]
      },
...
      "TeamImageUrlResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://bucket.s3.amazonaws.com/teams/uuid.png"
          }
        },
        "required": [
          "url"
        ]
      },
...
      "ImageUploadResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://cdn.example.com/players/uuid.jpg"
          }
        },
        "required": [
          "url"
        ]
      },
...
      "NewsImageUrlResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://cdn.example.com/news/abc.jpg",
            "description": "Yuklenen gorselin URL-si"
          }
        },
        "required": [
          "url"
        ]
      },
...
    }
  }
}